DjangoProjectBase/requirements.txt
django
Pillow
pandas
matplotlib
openai
python-dotenv
numpy
requests
